Extract and simplify functions for formatting ChartLine points
